refactor(ppp): split CLAS OSR continuity out of ppp_osr.cpp
- ppp_osr_internal.hpp: former anonymous-namespace helpers (QZSS held
  STEC cache, token suffix helpers) as inline functions in
  libgnss::ppp_osr_internal, shared by both TUs
- ppp_clas_osr_continuity.cpp: SSR orbit/clock boundary tracking, SIS
  change capture, atmosphere token/network selection and the computeOSR
  entry point (~1.7k lines)
- ppp_osr.cpp keeps the correction math (geometry compensation, bias
  lookup, phase-bias repair, dispersion compensation)

Verified: ctest 71/71, byte-identical rtk_positioning output on PPC
tokyo/run1, MADOCALIB parity suite 17/17.
